% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{Calc_mean_accuray_decrease}
\alias{Calc_mean_accuray_decrease}
\title{Calc_mean_accuracy_decrease}
\usage{
Calc_mean_accuray_decrease(impt_feat_list)
}
\arguments{
\item{impt_feat_list}{A list of dataframes containing the list of important features generated from running Run_RF_Pipeline.}
}
\value{
Returns a dataframe containing the mean, sd, max, and min decrease in accuracy for each feature over all of the random forest models that worked best during cross
validation.
}
\description{
Calc_mean_accuracy_decrease
}
