% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{generate_ROC_curve}
\alias{generate_ROC_curve}
\title{Generate_ROC_curve}
\usage{
generate_ROC_curve(RF_models, dataset, labels, title)
}
\arguments{
\item{RF_models}{A list containing caret RF_models. If using Run_RF_Pipeline this information will be contained within the fifth index of the returned list.}

\item{dataset}{The feature_table that was input into Run_RF_Pipeline}

\item{labels}{The actual labels for each sample(row) within the dataset}

\item{title}{The title of the plot}
}
\value{
Returns a ggplot object that plots the testing AUC of each of the best cross-validated random forest models generated when data is split into 
test and cross-validation datasets. Note that the redline represents the mean value for the senstivity and specificity at each step tested during ROC calculation
}
\description{
Generate_ROC_curve
}
