% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{remove_rare}
\alias{remove_rare}
\title{remove_rare}
\usage{
remove_rare(table, cutoff_pro)
}
\arguments{
\item{table}{The feature table to be filtered, (note rownames=features, columns=samples)}

\item{cutoff_pro}{The precent of samples a feature must be found in in order to be kept.}
}
\value{
A new table with any features that were below the cutoff proportion removed.
}
\description{
remove_rare
}
