% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{rf_classification_pipeline}
\alias{rf_classification_pipeline}
\title{Random Forest Classifcation pipeline}
\usage{
rf_classification_pipeline(
  feature_table,
  classes,
  metric = "ROC",
  ntree = 999,
  nmtry = 7,
  sampling = NULL,
  nfolds = 5,
  ncrossrepeats = 10,
  pro = 0.8,
  SEED = 1995
)
}
\arguments{
\item{feature_table}{The feature table that contains the information to be input into the random forest classifier. Note that this
table should not include information about the classes that are being predicted.}

\item{classes}{A vector that represents the classes that each sample (row) in the feature table represents. This can be coded as
Case (level 1 factor) and control (level 2 factor). Make sure the factor levels are correct with using AUPRC or results will not always be correct.}

\item{metric}{A string that indicates whether the pipeline should use AUROC or AUPRC. For AUROC set metric="ROC". For AUPRC set metric="PR".
Defaults to "ROC".}

\item{ntree}{An integer that represents the number of trees that you want to use during randoom forest construction. Defaults to 999.}

\item{nmtry}{An integer representing the number of different mtry values that you want to test during cross validation. The values of mtry to test
is calculated as follows:  mtry <- round(seq(1, number_of_features/3, length=nmtry)). Defaults to 7.}

\item{sampling}{A string indicating that type of sampling that should be done incase of inbalanced class designs. Options include: "up", "down" "SMOTE" and NULL.}

\item{nfolds}{An integer that represents the number of folds to used during cross validation. Defaults to 5.}

\item{ncrossrepeats}{An integer that represents the number of times to run cross validation on k folds. Defaults to 10.}

\item{pro}{The proporition of samples that should be used for training versus testing during cross validation. Defaults to 0.8}

\item{SEED}{The random seed used to split the samples during cross validation. Defaults to 1995.}
}
\description{
This function should not be run alone. You should use Run_RF_pipeline function to run the main pipeline for this package.
}
