% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{rf_regression_pipeline}
\alias{rf_regression_pipeline}
\title{rf_regression_pipeline}
\usage{
rf_regression_pipeline(feature_table, actual, SEED = 1995, sampling = NULL)
}
\arguments{
\item{feature_table}{The input feature table that contains the features (columns) and samples (rows) that should be used for model training and cross validation.}

\item{actual}{A vector containing the actual values for the parameter that is being regressed.}

\item{SEED}{An integer representing the random seed to split samples during cross validation.}
}
\description{
Function to run a random forest regression cross validation pipeline. Should not be run by itself. Currently a work in progress.
}
