% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{Run_RF_Pipeline}
\alias{Run_RF_Pipeline}
\title{Run_RF_Pipeline}
\usage{
Run_RF_Pipeline(
  feature_table,
  classes,
  metric = "ROC",
  sampling = NULL,
  repeats = 10,
  path,
  nmtry = 6,
  ntree = 1001,
  nfolds = 3,
  ncrossrepeats = 10,
  pro = 0.8,
  list_of_seeds
)
}
\arguments{
\item{feature_table}{The feature table that contains the information to be input into the random forest classifier. Note that this
table should not include information about the classes that are being predicted.}

\item{classes}{A vector that represents the classes that each sample (row) in the feature table represents. This can be coded as
Case (level 1 factor) and control (level 2 factor). Make sure the factor levels are correct with using AUPRC or results will not always be correct.}

\item{metric}{A string that indicates whether the pipeline should use AUROC or AUPRC. For AUROC set metric="ROC". For AUPRC set metric="PR".
Defaults to "ROC".}

\item{sampling}{A string indicating that type of sampling that should be done incase of inbalanced class designs. Options include: "up", "down" "SMOTE" and NULL.}

\item{repeats}{The number of times data should be split into testing and cross-validation datasets.}

\item{path}{A string representing the PATH were output files should be saved.}

\item{nmtry}{An integer representing the number of different mtry values that you want to test during cross validation. The values of mtry to test
is calculated as follows:  mtry <- round(seq(1, number_of_features/3, length=nmtry)). Defaults to 7.}

\item{ntree}{An integer that represents the number of trees that you want to use during randoom forest construction. Defaults to 1001.}

\item{nfolds}{An integer that represents the number of folds to used during cross validation. Defaults to 3.}

\item{ncrossrepeats}{An integer that represents the number of times to run cross validation on k folds. Defaults to 10.}

\item{pro}{The proporition of samples that should be used for training versus testing during cross validation. Defaults to 0.8}

\item{list_of_seeds}{A vector containing a number of seeds that should be equal to the number of repeats.}

\item{SEED}{The random seed used to split the samples during cross validation. Defaults to 1995.}
}
\value{
This function returns a list with the following characteristics:
"Object[[1]] contains all the median cross validation AUCS from each data split using the best mtry value"
"Object[[2]] contains all the test AUC values from each data split"
"Object[[3]] contains all the tested mtry values and the median ROC for each from each data split"
"Object[[4]] contains the list of important features from the best model selected from each data split"
"Object[[5]] contains each caret random forest model from each data split"
"This function will also write a csv with cross validation AUCS and test AUCS, to the given path as well as an RDS file that contains the resulting object from this function"
}
\description{
Run_RF_Pipeline
}
