% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{Run_RF_Regression_Pipeline}
\alias{Run_RF_Regression_Pipeline}
\title{Run_RF_Regression_Pipeline}
\usage{
Run_RF_Regression_Pipeline(
  feature_table,
  actual,
  sampling = NULL,
  repeats,
  path,
  list_of_seeds
)
}
\arguments{
\item{feature_table}{A feature table containing the samples (rows) and the features (columns) to run random forest regression on. Note that this table should
not include the value that is trying to be predicted}

\item{actual}{A vector containing the actual values for the value that this trying to be predectied.}

\item{sampling}{The sampling technique to use during cross validation. Defaults to NULL.}

\item{repeats}{The number of data splits that should occur between testing data and cross validation data.}

\item{path}{The path that the output should be saved to.}

\item{list_of_seeds}{A list of seeds equal to the length of repeats that is used for each random data split.}
}
\value{
Returns a list containing the following:
"Object[[1]] contains all the median cross validation RMSE from each data split using the best mtry value"
"Object[[2]] contains all the test RMSE values from each data split"
"Object[[3]] contains all the tested mtry values and the median RMSE from each from each data split"
"Object[[4]] contains the list of important features from the best model selected from each data split"
"Object[[5]] contains each caret random forest model from each data split"
"This function will also write a csv with cross validation RMSE and test RMSE, to the given path as well as an RDS file that contains the resulting object from this function"
}
\description{
Run_RF_Regression_Pipeline
}
