% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RF_Utilities.R
\name{set_cores}
\alias{set_cores}
\title{Set number of cores to run random forest pipeline on}
\usage{
set_cores(x)
}
\arguments{
\item{x}{Number of cores to run pipeline. Defaults to 1.}
}
\description{
Set number of cores to run random forest pipeline on
}
